Restructuring using MVC Pattern & Separate multi class in one java class file
Restructuring :  All source
Separate Class : PartitureSelectedListener.java
		 HNode.java
		 RBNode.java
		 Tree.java

At common GUI programming, avoid tight coupling between view and controller,
Apply Model View Controller Pattern
