+++ src/Model/HNode.java
@@ -1,0 +1,14 @@
+package src.Model;
+
+class HNode {
+    HNode left,right,p;
+    
+    Comparable key;
+    Object value;
+    
+    public HNode(Comparable key, Object value){
+        this.key=key;
+        this.value=value;
+        this.left=this.right=this.p=null;
+    }
+}
