+++ src/Model/RBNode.java
@@ -1,0 +1,11 @@
+package src.Model;
+
+class RBNode extends HNode{
+    RBNode left,right,p;
+    short color;
+    
+    public RBNode(Comparable key, Object data){
+        super(key,data);
+        this.color=RBTree.BLACK;
+    }
+}
